chore: increase sidebar width feat: demo mode - don't start services.

+++ cmd/main.go
@@ -48,6 +48,7 @@ var (
 	ctx         = context.Background()
 	appName     = "libredesk"
 	frontendDir = "frontend/dist"
+	envDemo     = "demo"
 
 	// Injected at build time.
 	buildString = ""
@@ -168,16 +169,18 @@ func main() {
 		conversation                = initConversations(i18n, sla, status, priority, wsHub, notifier, db, inbox, user, team, media, settings, csat, automation, template)
 		autoassigner                = initAutoAssigner(team, user, conversation)
 	)
-
 	automation.SetConversationStore(conversation)
-	startInboxes(ctx, inbox, conversation)
-	go automation.Run(ctx, automationWorkers)
-	go autoassigner.Run(ctx, autoAssignInterval)
-	go conversation.Run(ctx, messageIncomingQWorkers, messageOutgoingQWorkers, messageOutgoingScanInterval)
-	go conversation.RunUnsnoozer(ctx, unsnoozeInterval)
-	go notifier.Run(ctx)
-	go sla.Run(ctx, slaEvaluationInterval)
-	go media.DeleteUnlinkedMedia(ctx)
+
+	if ko.MustString("app.env") != envDemo {
+		startInboxes(ctx, inbox, conversation)
+		go automation.Run(ctx, automationWorkers)
+		go autoassigner.Run(ctx, autoAssignInterval)
+		go conversation.Run(ctx, messageIncomingQWorkers, messageOutgoingQWorkers, messageOutgoingScanInterval)
+		go conversation.RunUnsnoozer(ctx, unsnoozeInterval)
+		go notifier.Run(ctx)
+		go sla.Run(ctx, slaEvaluationInterval)
+		go media.DeleteUnlinkedMedia(ctx)
+	}
 
 	var app = &App{
 		lo:            lo,
@@ -224,6 +227,7 @@ func main() {
 	}
 
 	go func() {
+		colorlog.Green("Starting HTTP server in %s mode", ko.String("app.env"))
 		colorlog.Green("Server started at %s", ko.String("app.server.address"))
 		if ko.String("server.socket") != "" {
 			colorlog.Green("Unix socket created at %s", ko.String("server.socket"))
